// basketstats-app/src/app/models/players-summary-stats-of-match.ts
import {Player} from "./player";

export class PlayersSummaryStatsOfMatch {

  player: Player;

  id: number;

  pts: number;

  pm2: number;

  pa2: number;

  pm3: number;

  pa3: number;

  ftm: number;

  fta: number;

  ast: number;

  blkm: number;

  blkg: number;

  offr: number;

  defr: number;

  tov: number;

  stl: number;

  pf: number;

  fd: number;

  eff: number;

}